Restructure Python script into one
